battlenet v0.0.3 Battlenet.Auth

Provides functions for retrieving access tokens and authorizing OAuth clients using the Battlenet API.

Functions

authorize_url()

Used to get player authorization for you to access certain resources of theirs, like a list of World of Warcraft characters. For an authorization code request, you need to either redirect the browser or open a new window to this url.

Example

iex> Battlenet.Auth.authorize_url()
"https://US.battle.net/oauth/authorize?client_id=1234abcde
&redirect_uri=https://localhost.test/auth/callback&response_type=code"

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

clear_token()

Clears any stored access token obtained using client credentials.

Example

iex(1)> Battlenet.Auth.token()
"zevtpr9bt65g8dt4berkyp99"
iex(2)> Battlenet.Auth.clear_token()
:ok
iex(3)> Battlenet.Auth.token()
"eaf7dqe6t93sbfzh93v89kyr"

token()

Request access token that you can then use with other APIs. The token is validated on each successive token request. The token is saved in a separate process on process failure and fetched on re-initialization. If a token is expired, or was created by a different client id, a new token is fetched.

Example

iex> Battlenet.Auth.token()
"zevtpr9bt65g8dt4berkyp99"

token(code)

Turn in an authorization code for access tokens that you can then use with other APIs. This token is not persisted, and a new token will be fetched on each call.

Parameters

code: Authorization code that represents the user’s agreement to allow you data based on the scope.

Example

iex> Battlenet.Auth.token("123abc")
zevtpr9bt65g8dt4berkyp99"