battlenet v0.0.3 Battlenet.Auth.AccessToken

Provides model for access token and functions for verifying that the token is valid.

Functions

is_valid?(access_token)

Validates the access token string or struct

Parameters

access_token: the access token string or struct to be validated

Examples

iex(1)> token = Battlenet.Auth.token()
"zu5pu7ss3575nxe7zm4vupup"
iex(2)> Battlenet.Auth.AccessToken.is_valid?(token)
true
iex(3)> token_struct = %Battlenet.Auth.AccessToken{ access_token: token }
%Battlenet.Auth.AccessToken{
  access_token: "u5pu7ss3575nxe7zm4vupup",
  expires_in: nil,
  token_type: nil
}
iex(4)> Battlenet.Auth.AccessToken.is_valid?(token_struct)
true