battlenet v0.0.3 Battlenet.Resource

Provides generic functions for interacting with the Battlenet REST API

Functions

get(path)

Retrieve information for the provided path using a fetched and stored access token

Parameters

path: the url path for the desired resource.

Example

Battlenet.Resource.get("data/d3/season/")
{:ok,
%{
  "_links" => %{
    "self" => %{
      "href" => "https://us.api.battle.net/data/d3/season/?namespace=2-6-US"
    }
  },
  "current_season" => 13,
...

get(path, cast)

Retrieve information for the provided path and cast into provided map/struct using a fetched and stored access token.

Parameters

path: the url path for the desired resource

cast: a map or struct to cast the response as

get_with_api_key(path)

Retrieve information for the provided path using an api key.

Parameters

path: the url path for the desired resource

get_with_api_key(path, cast)

Retrieve information for the provided path using an api and casting into the provided map/struct

Parameters

path: the url path for the desired resource

cast: a map or struct to cast the response as

get_with_token(path, access_token)

Retrieve information for the provided path using the provided access token

Parameters

path: the url path for the desired resource

access_token: the access token string to use

get_with_token(path, access_token, cast)

Retrieve information for the provided path using the provided access token into provided map/struct.

Parameters

path: the url path for the desired resource

access_token: the access token or acces token string string to use

cast: a map or struct to cast the response as